+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this "vscode-msr" extension will be documented in this file.
 
+## [2.1.73]
+- Supported custom initialization command line by new config `msr.xxx.postInitTerminalCommandLine`.
+- Auto detect old git and skip `--recurse-submodules` for `git ls-files` command + shortcuts.
+- Added overriding of `C#` custom search command by config `msr.cs.myFindOrReplaceSelectedTextCommand`.
+- Added alias `clear-msr-env` + Added `trust-exe` on Windows.
+
 ## [2.1.72]
 - Improved custom search command `msr.xxx.myFindOrReplaceSelectedTextCommand`:
   - Supported short `%Skip_Git_Paths%` which not exported as a tmp env var.